refactor igv
