+++ malariagen_data/ag3.py
@@ -581,30 +581,6 @@ class Ag3(AnophelesDataResource):
             chunks=chunks,
         )
 
-    def _view_alignments_add_site_filters_tracks(
-        self, *, contig, visibility_window, tracks
-    ):
-        debug = self._log.debug
-
-        for site_mask in self.site_mask_ids:
-            site_filters_vcf_url = f"gs://vo_agam_release/v3/site_filters/{self._site_filters_analysis}/vcf/{site_mask}/{contig}_sitefilters.vcf.gz"  # noqa
-            debug(site_filters_vcf_url)
-            track_config = {
-                "name": f"Filters - {site_mask}",
-                "url": site_filters_vcf_url,
-                "indexURL": f"{site_filters_vcf_url}.tbi",
-                "format": "vcf",
-                "type": "variant",
-                "visibilityWindow": visibility_window,  # bp
-                "height": 30,
-                "colorBy": "FILTER",
-                "colorTable": {
-                    "PASS": "#00cc96",
-                    "*": "#ef553b",
-                },
-            }
-            tracks.append(track_config)
-
     def _results_cache_add_analysis_params(self, params):
         super()._results_cache_add_analysis_params(params)
         # override parent class to add AIM analysis
