+++ malariagen_data/anopheles.py
@@ -11,7 +11,6 @@ import bokeh.models
 import bokeh.palettes
 import bokeh.plotting
 import dask
-import igv_notebook
 import numba
 import numpy as np
 import pandas as pd
@@ -48,6 +47,7 @@ from .anoph.cnv_data import AnophelesCnvData
 from .anoph.genome_features import AnophelesGenomeFeaturesData
 from .anoph.genome_sequence import AnophelesGenomeSequenceData
 from .anoph.hap_data import AnophelesHapData, hap_params
+from .anoph.igv import AnophelesIgv
 from .anoph.pca import AnophelesPca
 from .anoph.sample_metadata import AnophelesSampleMetadata
 from .anoph.snp_data import AnophelesSnpData
@@ -99,6 +99,7 @@ DEFAULT_MAX_COVERAGE_VARIANCE = 0.2
 # noinspection PyCallingNonCallable
 class AnophelesDataResource(
     AnophelesPca,
+    AnophelesIgv,
     AnophelesCnvData,
     AnophelesAimData,
     AnophelesHapData,
@@ -216,15 +217,6 @@ class AnophelesDataResource(
     def _transcript_to_gene_name(self, transcript):
         # children may have different manual overrides.
         raise NotImplementedError("Must override _transcript_to_gene_name")
-
-    @abstractmethod
-    def _view_alignments_add_site_filters_tracks(
-        self, *, contig, visibility_window, tracks
-    ):
-        # default implementation, do nothing
-        raise NotImplementedError(
-            "Must override _view_alignments_add_site_filters_tracks"
-        )
 
     @check_types
     @doc(
@@ -796,130 +788,6 @@ class AnophelesDataResource(
         ann.get_effects(transcript=transcript, variants=df_snps)
 
         return df_snps
-
-    @check_types
-    @doc(
-        summary="Create an IGV browser and inject into the current notebook.",
-        parameters=dict(
-            tracks="Configuration for any additional tracks.",
-        ),
-        returns="IGV browser.",
-    )
-    def igv(
-        self, region: base_params.region, tracks: Optional[List] = None
-    ) -> igv_notebook.Browser:
-        debug = self._log.debug
-
-        debug("resolve region")
-        region_prepped: Region = parse_single_region(self, region)
-        del region
-
-        config = {
-            "reference": {
-                "id": self._genome_ref_id,
-                "name": self._genome_ref_name,
-                "fastaURL": f"{self._gcs_url}{self._genome_fasta_path}",
-                "indexURL": f"{self._gcs_url}{self._genome_fai_path}",
-                "tracks": [
-                    {
-                        "name": "Genes",
-                        "type": "annotation",
-                        "format": "gff3",
-                        "url": f"{self._gcs_url}{self._geneset_gff3_path}",
-                        "indexed": False,
-                    }
-                ],
-            },
-            "locus": str(region_prepped),
-        }
-        if tracks:
-            config["tracks"] = tracks
-
-        debug(config)
-
-        igv_notebook.init()
-        browser = igv_notebook.Browser(config)
-
-        return browser
-
-    @check_types
-    @doc(
-        summary="""
-            Launch IGV and view sequence read alignments and SNP genotypes from
-            the given sample.
-        """,
-        parameters=dict(
-            sample="Sample identifier.",
-            visibility_window="""
-                Zoom level in base pairs at which alignment and SNP data will become
-                visible.
-            """,
-        ),
-    )
-    def view_alignments(
-        self,
-        region: base_params.region,
-        sample: str,
-        visibility_window: int = 20_000,
-    ):
-        debug = self._log.debug
-
-        debug("look up sample set for sample")
-        sample_rec = self.sample_metadata().set_index("sample_id").loc[sample]
-        sample_set = sample_rec["sample_set"]
-
-        debug("load data catalog")
-        df_cat = self.wgs_data_catalog(sample_set=sample_set)
-
-        debug("locate record for sample")
-        cat_rec = df_cat.set_index("sample_id").loc[sample]
-        bam_url = cat_rec["alignments_bam"]
-        vcf_url = cat_rec["snp_genotypes_vcf"]
-        debug(bam_url)
-        debug(vcf_url)
-
-        debug("parse region")
-        resolved_region: Region = parse_single_region(self, region)
-        del region
-        contig = resolved_region.contig
-
-        # begin creating tracks
-        tracks: List[Dict] = []
-
-        # https://github.com/igvteam/igv-notebook/issues/3 -- resolved now
-        debug("set up site filters tracks")
-        self._view_alignments_add_site_filters_tracks(
-            contig=contig, visibility_window=visibility_window, tracks=tracks
-        )
-
-        debug("add SNPs track")
-        tracks.append(
-            {
-                "name": "SNPs",
-                "url": vcf_url,
-                "indexURL": f"{vcf_url}.tbi",
-                "format": "vcf",
-                "type": "variant",
-                "visibilityWindow": visibility_window,  # bp
-                "height": 50,
-            }
-        )
-
-        debug("add alignments track")
-        tracks.append(
-            {
-                "name": "Alignments",
-                "url": bam_url,
-                "indexURL": f"{bam_url}.bai",
-                "format": "bam",
-                "type": "alignment",
-                "visibilityWindow": visibility_window,  # bp
-                "height": 500,
-            }
-        )
-
-        debug("create IGV browser")
-        self.igv(region=resolved_region, tracks=tracks)
 
     def _cohort_alt_allele_counts_melt(self, gt, indices, max_allele):
         ac_alt_melt, an = self._cohort_alt_allele_counts_melt_kernel(
